Image to text/requirement.txt
Flask
pytesseract
pillow
opencv-python
numpy
